// src/types.ts
export type Event = {
  committee: string;
  cover: string;
  description: string;
  end: string;
  start: string;
  location: string;
  organization: string;
  pointValue: number;
  requiresStaff: boolean;
  staffPointBonus: number;
  title: string;
  thumbnail: string;
  uuid: string;
  // checked: boolean;
};